requirements.txt
alembic
bcrypt
blinker
click
colorama
Flask==2.3.3
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
flask-wtf==1.1.1
Flask-WTF
greenlet
itsdangerous
Jinja2
Mako
MarkupSafe
python-dotenv
SQLAlchemy
typing_extensions
Werkzeug
WTForms
firebase-admin
flask-cors
pyrebase4
requests
setuptools